W5 hw2 add controller
