+++ homeworks/week5/hw2/controller/route.js
@@ -1,0 +1,13 @@
+module.exports = {
+  index: (req, res) =>{
+    res.render('index');
+  },
+  
+  login: (req, res) => {
+    res.render('login');
+  },
+
+  register: (req, res) => {
+    res.render('register');
+  }
+}
